Library documentation

TNRKit.ATRG — Type

mutable struct ATRG <: TNRScheme

Anisotropic Tensor Renormalization Group

Constructors

ATRG(T [, finalize=finalize!])

Running the algorithm

run!(::ATRG, trunc::TensorKit.TruncationSheme, stop::Stopcrit[, finalize_beginning=true, verbosity=1])

Each step rescales the lattice by a (linear) factor of √2

verbosity levels

0: No output

1: Print information at start and end of the algorithm

2: Print information at each step

Fields

T::TensorKit.TensorMap

finalize!::Function

References

Adachi et. al. Phys. Rev. B 102 (2020)

TNRKit.ATRG_3D — Type

mutable struct ATRG_3D <: TNRScheme

3D Anisotropic Tensor Renormalization Group

Constructors

ATRG_3D(T [, finalize=finalize!])

Running the algorithm

run!(::ATRG_3D, trunc::TensorKit.TruncationSheme, stop::Stopcrit[, finalize_beginning=true,verbosity=1])

Each step rescales the lattice by a (linear) factor of 2

verbosity levels

0: No output

1: Print information at start and end of the algorithm

2: Print information at each step

Fields

T::TensorKit.TensorMap

finalize!::Function

References

Adachi et. al. Phys. Rev. B 102 (2020)

TNRKit.BTRG — Type

mutable struct BTRG <: TNRScheme

Bond-weighted Tensor Renormalization Group

Constructors

BTRG(T [, k=-1/2, finalize=finalize!])

Running the algorithm

run!(::BTRG, trunc::TensorKit.TruncationSheme, stop::Stopcrit[, finalize_beginning=true, verbosity=1])

Each step rescales the lattice by a (linear) factor of √2

verbosity levels

0: No output

1: Print information at start and end of the algorithm

2: Print information at each step

Fields

T::TensorKit.TensorMap

S1::TensorKit.TensorMap

S2::TensorKit.TensorMap

k::Float64

finalize!::Function

References

Adachi et. al. Phys. Rev. B 105 (2022)

TNRKit.HOTRG — Type

mutable struct HOTRG <: TNRScheme

Higher-Order Tensor Renormalization Group

Constructors

HOTRG(T [, finalize=finalize!])

Running the algorithm

run!(::HOTRG, trunc::TensorKit.TruncationSheme, stop::Stopcrit[, finalize_beginning=true, verbosity=1])

Each step rescales the lattice by a (linear) factor of 2

verbosity levels

0: No output

1: Print information at start and end of the algorithm

2: Print information at each step

Fields

T::TensorKit.TensorMap

finalize!::Function

References

Xie et. al. Phys. Rev. B 86 (2012)

TNRKit.LoopTNR — Type

mutable struct LoopTNR <: TNRScheme

Loop Optimization for Tensor Network Renormalization

Constructors

LoopTNR(T [, finalize=finalize!])
LoopTNR(TA, TB, [, finalize=finalize!])

Running the algorithm

run!(::LoopTNR, trunc::TensorKit.TruncationScheme, truncentanglement::TensorKit.TruncationScheme, criterion::stopcrit,
          entanglement_criterion::stopcrit, loop_criterion::stopcrit[, finalize_beginning=true, verbosity=1])

run!(::LoopTNR, trscheme::TensorKit.TruncationScheme, criterion::stopcrit[, finalize_beginning=true, verbosity=1])

Fields

TA::TensorKit.TensorMap

TB::TensorKit.TensorMap

finalize!::Function

References

Yang et. al. Phys. Rev. Letters 118 (2017)

TNRKit.SLoopTNR — Type

mutable struct SLoopTNR <: TNRScheme

c4 & inversion symmetric Loop Optimization for Tensor Network Renormalization

Constructors

SLoopTNR(T [, finalize=finalize!])
SLoopTNR(TA, TB, [, finalize=finalize!])

Running the algorithm

run!(::SLoopTNR, trscheme::TensorKit.TruncationScheme,
          criterion::TNRKit.stopcrit[, finalize_beginning=true, oneloop=true,
          verbosity=1])

oneloop=true will use disentangled tensors as a starting guess for the optimization.

Fields

T::TensorKit.TensorMap

gradalg::OptimKit.LBFGS

finalize!::Function

References

Yang et. al. Phys. Rev. Letters 118 (2017) (Fig. S6)

TNRKit.TRG — Type

mutable struct TRG <: TNRScheme

Tensor Renormalization Group

Constructors

TRG(T [, finalize=finalize!])

Running the algorithm

run!(::TRG, trunc::TensorKit.TruncationSheme, stop::Stopcrit[, finalize_beginning=true, verbosity=1])

Each step rescales the lattice by a (linear) factor of √2

verbosity levels

0: No output

1: Print information at start and end of the algorithm

2: Print information at each step

Fields

T::TensorKit.TensorMap: central tensor

finalize!::Function: finalization function

References

Levin & Nave Phys. Rev. Letters 99(12) (2007)

TNRKit.central_charge — Method

central_charge(scheme::TNRScheme, n::Number)

Get the central charge given the current state of a TNRScheme and the previous normalization factor n

TNRKit.classical_clock — Method

classical_clock(q, β)

Constructs the partition function tensor for the classical clock model with q states and a given inverse temperature β.

TNRKit.classical_ising — Method

classical_ising(β; h)

Constructs the partition function tensor for a 2D square lattice for the classical Ising model with a given inverse temperature β and external magnetic field h.

Examples

    classical_ising() # Default inverse temperature is `ising_βc`
    classical_ising(0.5; h = 1.0) # Custom inverse temperature and magnetic field.

Info

When calculating the free energy with free_energy(), set the initial_size keyword argument to 2.0. The initial lattice holds 2 spins.

See also: classical_ising_symmetric, classical_ising_symmetric_3D, classical_ising_3D.

TNRKit.classical_ising_3D — Method

classical_ising_3D(β; J)

Constructs the partition function tensor for a 3D cubic lattice for the classical Ising model with a given inverse temperature β and coupling constant J (by default J = 1.0).

Examples

    classical_ising_3D() # Default inverse temperature is `ising_βc_3D`, coupling constant is `J = 1.0`.
    classical_ising_3D(0.5; J = 1.0) # Custom inverse temperature and coupling constant.

See also: classical_ising_symmetric_3D, classical_ising, classical_ising_symmetric.

TNRKit.classical_ising_symmetric — Method

classical_ising_symmetric(β)

Constructs the partition function tensor for a symmetric 2D square lattice for the classical Ising model with a given inverse temperature β.

This tensor has explicit ℤ₂ symmetry on each of it spaces.

Examples

    classical_ising_symmetric() # Default inverse temperature is `ising_βc`
    classical_ising_symmetric(0.5) # Custom inverse temperature.

See also: classical_ising, classical_ising_symmetric_3D, classical_ising_3D.

TNRKit.classical_ising_symmetric_3D — Method

classical_ising_symmetric_3D(β)

Constructs the partition function tensor for a symmetric 3D cubic lattice for the classical Ising model with a given inverse temperature β.

This tensor has explicit ℤ₂ symmetry on each of its spaces.

Examples

    classical_ising_symmetric_3D() # Default inverse temperature is `ising_βc_3D`
    classical_ising_symmetric_3D(0.5) # Custom inverse temperature.

See also: classical_ising_3D, classical_ising, classical_ising_symmetric.

TNRKit.classical_potts — Method

classical_potts(q, β)

Constructs the partition function tensor for the classical Potts model with q states and a given inverse temperature β.

Examples

    classical_potts(3) # Default inverse temperature is `potts_βc(3)`
    classical_potts(3, 0.5) # Custom inverse temperature.

See also: classical_potts_symmetric, potts_βc.

TNRKit.classical_potts_symmetric — Method

classical_potts_symmetric(q, β)

Constructs the partition function tensor for a symmetric Potts model with q states and a given inverse temperature β.

This tensor has explicit ℤq symmetry on each of its spaces.

Examples

    classical_potts_symmetric(3) # Default inverse temperature is `potts_βc(3)`
    classical_potts_symmetric(3, 0.5) # Custom inverse temperature.

See also: classical_potts, potts_βc.

TNRKit.free_energy — Method

free_energy(data, β; scalefactor, initial_size)

Takes a vector of normalization factors data and a given inverse temperature β and computes the free energy.

Info

The scalefactor should be set to the rescaling factor of the area of the tensor network after each iteration of the TNR algorithm.

The initial_size should be set to the intial size of the physical lattice, which is typically 1.0 (or 2.0 for the 2D non-symmetric classical Ising model).

TNRKit.gross_neveu_start — Method

gross_neveu_start(μ, m, g)

Constructs the partition function tensor for the Gross-Neveu model with given parameters μ, m, and g.

References

Akiyama et. al. J. Phys.: Condens. Matter 36 (2024) 343002

TNRKit.potts_βc — Method

potts_βc(q)

returns the inverse critical temperature for the classical q-state Potts model on a 2D square lattice.

See also: classical_potts, classical_potts_symmetric.

TNRKit.sixvertex — Function

sixvertex(elt::Type{<:Number} = Float64, symmetry::Type{<:Sector} = Trivial; a = 1.0, b = 1.0, c = 1.0)
sixvertex(symmetry::Type{<:Sector}; kwargs...)

Constructs the partition function tensor for the six-vertex model with a given symmetry type and coupling constants a, b, and c.

Defaults

- elt: Float64
- symmetry: Trivial
- a: 1.0
- b: 1.0
- c: 1.0

Examples

    sixvertex() # Default symmetry is `Trivial`, coupling constants are `a = 1.0`, `b = 1.0`, `c = 1.0`.
    sixvertex(CU1Irrep) # CU1 symmetry with default coupling constants.
    sixvertex(ComplexF64, U1Irrep; a = 2.0, b = 3.0, c = 4.0) # U1 symmetry with custom coupling constants and element type.