Library documentation

TNRKit.ATRG — Type

mutable struct ATRG <: TNRScheme

Anisotropic Tensor Renormalization Group

Constructors

ATRG(T [, finalize=finalize!])

Running the algorithm

run!(::ATRG, trunc::TensorKit.TruncationSheme, stop::Stopcrit[, finalize_beginning=true, verbosity=1])

Each step rescales the lattice by a (linear) factor of √2

verbosity levels

0: No output

1: Print information at start and end of the algorithm

2: Print information at each step

Fields

T::TensorKit.TensorMap

finalize!::Function

References

Adachi et. al. Phys. Rev. B 102 (2020)

TNRKit.ATRG_3D — Type

mutable struct ATRG_3D <: TNRScheme

3D Anisotropic Tensor Renormalization Group

Constructors

ATRG_3D(T [, finalize=finalize!])

Running the algorithm

run!(::ATRG_3D, trunc::TensorKit.TruncationSheme, stop::Stopcrit[, finalize_beginning=true,verbosity=1])

Each step rescales the lattice by a (linear) factor of 2

verbosity levels

0: No output

1: Print information at start and end of the algorithm

2: Print information at each step

Fields

T::TensorKit.TensorMap

finalize!::Function

References

Adachi et. al. Phys. Rev. B 102 (2020)

TNRKit.BTRG — Type

mutable struct BTRG <: TNRScheme

Bond-weighted Tensor Renormalization Group

Constructors

BTRG(T [, k=-1/2, finalize=finalize!])

Running the algorithm

run!(::BTRG, trunc::TensorKit.TruncationSheme, stop::Stopcrit[, finalize_beginning=true, verbosity=1])

Each step rescales the lattice by a (linear) factor of √2

verbosity levels

0: No output

1: Print information at start and end of the algorithm

2: Print information at each step

Fields

T::TensorKit.TensorMap

S1::TensorKit.TensorMap

S2::TensorKit.TensorMap

k::Float64

finalize!::Function

References

Adachi et. al. Phys. Rev. B 105 (2022)

TNRKit.HOTRG — Type

mutable struct HOTRG <: TNRScheme

Higher-Order Tensor Renormalization Group

Constructors

HOTRG(T [, finalize=finalize!])

Running the algorithm

run!(::HOTRG, trunc::TensorKit.TruncationSheme, stop::Stopcrit[, finalize_beginning=true, verbosity=1])

Each step rescales the lattice by a (linear) factor of 2

verbosity levels

0: No output

1: Print information at start and end of the algorithm

2: Print information at each step

Fields

T::TensorKit.TensorMap

finalize!::Function

References

Xie et. al. Phys. Rev. B 86 (2012)

TNRKit.LoopTNR — Type

mutable struct LoopTNR <: TNRScheme

Loop Optimization for Tensor Network Renormalization

Constructors

LoopTNR(T [, finalize=finalize!])
LoopTNR(TA, TB, [, finalize=finalize!])

Running the algorithm

run!(::LoopTNR, trunc::TensorKit.TruncationScheme, truncentanglement::TensorKit.TruncationScheme, criterion::stopcrit,
          entanglement_criterion::stopcrit, loop_criterion::stopcrit[, finalize_beginning=true, verbosity=1])

run!(::LoopTNR, trscheme::TensorKit.TruncationScheme, criterion::stopcrit[, finalize_beginning=true, verbosity=1])

Fields

TA::TensorKit.TensorMap

TB::TensorKit.TensorMap

finalize!::Function

References

Yang et. al. Phys. Rev. Letters 118 (2017)

TNRKit.SLoopTNR — Type

mutable struct SLoopTNR <: TNRScheme

c4 & inversion symmetric Loop Optimization for Tensor Network Renormalization

Constructors

SLoopTNR(T [, finalize=finalize!])
SLoopTNR(TA, TB, [, finalize=finalize!])

Running the algorithm

run!(::SLoopTNR, trscheme::TensorKit.TruncationScheme,
          criterion::TNRKit.stopcrit[, finalize_beginning=true, oneloop=true,
          verbosity=1])

oneloop=true will use disentangled tensors as a starting guess for the optimization.

Fields

T::TensorKit.TensorMap

gradalg::OptimKit.LBFGS

finalize!::Function

References

Yang et. al. Phys. Rev. Letters 118 (2017) (Fig. S6)

TNRKit.TRG — Type

mutable struct TRG <: TNRScheme

Tensor Renormalization Group

Constructors

TRG(T [, finalize=finalize!])

Running the algorithm

run!(::TRG, trunc::TensorKit.TruncationSheme, stop::Stopcrit[, finalize_beginning=true, verbosity=1])

Each step rescales the lattice by a (linear) factor of √2

verbosity levels

0: No output

1: Print information at start and end of the algorithm

2: Print information at each step

Fields

T::TensorKit.TensorMap: central tensor

finalize!::Function: finalization function

References

Levin & Nave Phys. Rev. Letters 99(12) (2007)

TNRKit.central_charge — Method

central_charge(scheme::TNRScheme, n::Number)

Get the central charge given the current state of a TNRScheme and the previous normalization factor n