Library documentation

TNRKit.ATRG — Type

mutable struct ATRG{E, S, TT<:TensorKit.AbstractTensorMap{E, S, 2, 2}} <: TNRScheme{E, S}

Anisotropic Tensor Renormalization Group

Constructors

ATRG(T)

Running the algorithm

run!(::ATRG, trunc::TensorKit.TruncationSheme, stop::Stopcrit[, finalizer=default_Finalizer, finalize_beginning=true, verbosity=1])

Each step rescales the lattice by a (linear) factor of √2

verbosity levels

0: No output

1: Print information at start and end of the algorithm

2: Print information at each step

Fields

T::TensorKit.AbstractTensorMap{E, S, 2, 2} where {E, S}: Central tensor

References

Adachi et. al. Phys. Rev. B 102 (2020)

TNRKit.ATRG_3D — Type

mutable struct ATRG_3D{E, S, TT<:TensorKit.AbstractTensorMap{E, S, 2, 4}} <: TNRScheme{E, S}

3D Anisotropic Tensor Renormalization Group

Constructors

ATRG_3D(T)

Running the algorithm

run!(::ATRG_3D, trunc::TensorKit.TruncationSheme, stop::Stopcrit[, finalizer=defualt_Finalizer, finalize_beginning=true,verbosity=1])

Each step rescales the lattice by a (linear) factor of 2

verbosity levels

0: No output

1: Print information at start and end of the algorithm

2: Print information at each step

Fields

T::TensorKit.AbstractTensorMap{E, S, 2, 4} where {E, S}: Central tensor

References

Adachi et. al. Phys. Rev. B 102 (2020)

TNRKit.BTRG — Type

mutable struct BTRG{E, S, TT<:TensorKit.AbstractTensorMap{E, S, 2, 2}, BT<:TensorKit.AbstractTensorMap{E, S, 1, 1}} <: TNRScheme{E, S}

Bond-weighted Tensor Renormalization Group

Constructors

BTRG(T [, k=-1/2])

Running the algorithm

run!(::BTRG, trunc::TensorKit.TruncationSheme, stop::Stopcrit[, finalizer=default_Finalizer, finalize_beginning=true, verbosity=1])

Each step rescales the lattice by a (linear) factor of √2

verbosity levels

0: No output

1: Print information at start and end of the algorithm

2: Print information at each step

Fields

T::TensorKit.AbstractTensorMap{E, S, 2, 2} where {E, S}: Central tensor

S1::TensorKit.AbstractTensorMap{E, S, 1, 1} where {E, S}: Bond tensor on vertical bonds

S2::TensorKit.AbstractTensorMap{E, S, 1, 1} where {E, S}: Bond tensor on horizontal bonds

k::Float64: Bond weight exponent

References

Adachi et. al. Phys. Rev. B 105 (2022)

TNRKit.CTM — Type

        ┌───┐ ┌──┐ ┌───┐
        │Ctl├─┤Et├─┤Ctr│
        └─┬─┘ └┬─┘ └─┬─┘
        ┌─┼─┐ ┌┼─┐ ┌─┼─┐
        │El ├─┤T ├─┤Er │
        └─┬─┘ └┬─┘ └─┬─┘
        ┌─┼─┐ ┌┼─┐ ┌─┼─┐
        │Cbl├─┤Eb├─┤Cbr│
        └───┘ └──┘ └───┘

TNRKit.Finalizer — Type

struct Finalizer{E}

Finalizer for TNR schemes

Constructors

Finalizer(f!::Function, E::Type)

A Finalizer holds a function f! that is to be applied to a TNR scheme after each step of the algorithm (and at the beginning if specified by run!(;finalize_beginning=true), which is the default behavior). The type parameter E indicates the output type of f!, which is used to create an array of the correct type to hold the outputs.

TNRKit.HOTRG — Type

mutable struct HOTRG{E, S, TT<:TensorKit.AbstractTensorMap{E, S, 2, 2}} <: TNRScheme{E, S}

Higher-Order Tensor Renormalization Group

Constructors

HOTRG(T)

Running the algorithm

run!(::HOTRG, trunc::TensorKit.TruncationSheme, stop::Stopcrit[, finalizer=default_Finalizer, finalize_beginning=true, verbosity=1])

Each step rescales the lattice by a (linear) factor of 2

verbosity levels

0: No output

1: Print information at start and end of the algorithm

2: Print information at each step

Fields

T::TensorKit.AbstractTensorMap{E, S, 2, 2} where {E, S}: Central tensor

References

Xie et. al. Phys. Rev. B 86 (2012)

TNRKit.HOTRG_3D — Type

mutable struct HOTRG_3D{E, S, TT<:TensorKit.AbstractTensorMap{E, S, 2, 4}} <: TNRScheme{E, S}

3D Higher-Order Tensor Renormalization Group

Constructors

HOTRG_3D(T)

Running the algorithm

run!(::HOTRG_3D, trunc::TensorKit.TruncationSheme, stop::Stopcrit[, finalizer=default_Finalizer, finalize_beginning=true, verbosity=1])

Each step rescales the lattice by a (linear) factor of 2

verbosity levels

0: No output

1: Print information at start and end of the algorithm

2: Print information at each step

Fields

T::TensorKit.AbstractTensorMap{E, S, 2, 4} where {E, S}: Central tensor

References

Xie et. al. Phys. Rev. B 86 (2012)

TNRKit.ImpurityHOTRG — Type

mutable struct ImpurityHOTRG{E, S, TT<:TensorKit.AbstractTensorMap{E, S, 2, 2}} <: TNRScheme{E, S}

Single impurity method for Higher-Order Tensor Renormalization Group (for 2nd order)

Constructors

ImpurityHOTRG(T, T_imp_order1_1, T_imp_order1_2, T_imp_order2)

Running the algorithm

run!(::ImpurityHOTRG, trunc::TensorKit.TruncationSheme, stop::Stopcrit[, finalizer=ImpurityHOTRG_Finalizer, finalize_beginning=true, verbosity=1])

Each step rescales the lattice by a (linear) factor of 2

verbosity levels

0: No output

1: Print information at start and end of the algorithm

2: Print information at each step

Fields

T::TensorKit.AbstractTensorMap{E, S, 2, 2} where {E, S}: Central Tensor

T_imp_order1_1::TensorKit.AbstractTensorMap{E, S, 2, 2} where {E, S}: +1 Type first order impurity tensor

T_imp_order1_2::TensorKit.AbstractTensorMap{E, S, 2, 2} where {E, S}: -1 Type first order impurity tensor

T_imp_order2::TensorKit.AbstractTensorMap{E, S, 2, 2} where {E, S}: Second order impurity tensor

References

Morita et al 10.1016/j.cpc.2018.10.014 (2018)

TNRKit.ImpurityTRG — Type

mutable struct ImpurityTRG{E, S, TT<:TensorKit.AbstractTensorMap{E, S, 2, 2}} <: TNRScheme{E, S}

Impurity method for Tensor Renormalization Group

Constructors

ImpurityTRG(T, T_imp1, T_imp2, T_imp3, T_imp4)

Running the algorithm

run!(::ImpurityTRG, trunc::TensorKit.TruncationSheme, stop::Stopcrit[, finalizer=ImpurityTRG_Finalizer, finalize_beginning=true, verbosity=1])

Each step rescales the lattice by a (linear) factor of √2

verbosity levels

0: No output

1: Print information at start and end of the algorithm

2: Print information at each step

Fields

T::TensorKit.AbstractTensorMap{E, S, 2, 2} where {E, S}: Pure tensor

T_imp1::TensorKit.AbstractTensorMap{E, S, 2, 2} where {E, S}: Impurity tensor on lattice site 1

T_imp2::TensorKit.AbstractTensorMap{E, S, 2, 2} where {E, S}: Impurity tensor on lattice site 2

T_imp3::TensorKit.AbstractTensorMap{E, S, 2, 2} where {E, S}: Impurity tensor on lattice site 3

T_imp4::TensorKit.AbstractTensorMap{E, S, 2, 2} where {E, S}: Impurity tensor on lattice site 4

Illustration (with p the pure tensor, and the p fills in the rest of the lattice):

    p   p
    |   |  
p---1---2---p
    |   |
p---4---3---p
    |   |
    p   p

References

Kadoh et. al. 10.1007/JHEP05(2019)184 (2019)

Morita et. al. Phys. Rev. B 111 (2025)

TNRKit.LoopTNR — Type

mutable struct LoopTNR{E, S, TT<:TensorKit.AbstractTensorMap{E, S, 2, 2}} <: TNRScheme{E, S}

Loop Optimization for Tensor Network Renormalization

Constructors

LoopTNR(T)
LoopTNR(TA, TB)
LoopTNR(unitcell_2x2::Matrix{T})

Running the algorithm

run!(::LoopTNR, trunc::TensorKit.TruncationScheme, truncentanglement::TensorKit.TruncationScheme, criterion::stopcrit,
          entanglement_criterion::stopcrit, loop_criterion::stopcrit[, finalize_beginning=true, verbosity=1])

run!(::LoopTNR, trscheme::TensorKit.TruncationScheme, criterion::stopcrit[, finalizer=default_Finalizer, finalize_beginning=true, verbosity=1])

Fields

TA::TensorKit.AbstractTensorMap{E, S, 2, 2} where {E, S}: Central tensor on sublattice A

TB::TensorKit.AbstractTensorMap{E, S, 2, 2} where {E, S}: Central tensor on sublattice B

References

Yang et. al. Phys. Rev. Letters 118 (2017)

TNRKit.LoopTNR — Method

LoopTNR(
    unitcell_2x2::Matrix{T},
    loop_criterion::stopcrit,
    trunc::TensorKit.TruncationScheme,
    truncentanglement::TensorKit.TruncationScheme
) where {T <: AbstractTensorMap{<:Any, <:Any, 2, 2}}

Initialize LoopTNR using a network with 2 x 2 unit cell, by first performing one round of loop optimization to reduce the network to a bipartite one (without normalization).

TNRKit.SLoopTNR — Type

mutable struct SLoopTNR{E, S, TT<:TensorKit.AbstractTensorMap{E, S, 4, 0}} <: TNRScheme{E, S}

c4 & inversion symmetric Loop Optimization for Tensor Network Renormalization

Constructors

SLoopTNR(T)
SLoopTNR(TA, TB)

Running the algorithm

run!(::SLoopTNR, trscheme::TensorKit.TruncationScheme,
          criterion::TNRKit.stopcrit[, finalizer=default_Finalizer, finalize_beginning=true, oneloop=true,
          verbosity=1])

oneloop=true will use disentangled tensors as a starting guess for the optimization.

Fields

T::TensorKit.AbstractTensor{E, S, 4} where {E, S}: Central tensor

gradalg::OptimKit.LBFGS: Gradient optimization algorithm

References

Yang et. al. Phys. Rev. Letters 118 (2017) (Fig. S6)

TNRKit.Sublattice_CTM — Type

– CTMRG for A-B sublattice systems

┌────┐ ┌───┐ ┌───┐ ┌────┐ │Ctl1├──┤EtB├──┤EtA├──┤Ctr1│ └─┬──┘ └─┬─┘ └─┬─┘ └─┬──┘ ┌─┼──┐ ┌─┼─┐ ┌─┼─┐ ┌─┼──┐ │ElB ├──┤ A ├──┤ B ├──┤ErA │ └─┬──┘ └─┬─┘ └─┬─┘ └─┬──┘ ┌─┼──┐ ┌─┼─┐ ┌─┼─┐ ┌─┼──┐ │ElA ├──┤ B ├──┤ A ├──┤ErB │ └─┬──┘ └─┬─┘ └─┬─┘ └─┬──┘ ┌─┼──┐ ┌─┼─┐ ┌─┼─┐ ┌─┼──┐ │Cbl1├──┤EbA├──┤EbB├──┤Cbr1│ └────┘ └───┘ └───┘ └────┘

┌────┐ ┌───┐ ┌───┐ ┌────┐ │Ctl2├──┤EtA├──┤EtB├──┤Ctr2│ └─┬──┘ └─┬─┘ └─┬─┘ └─┬──┘ ┌─┼──┐ ┌─┼─┐ ┌─┼─┐ ┌─┼──┐ │ElA ├──┤ B ├──┤ A ├──┤ErB │ └─┬──┘ └─┬─┘ └─┬─┘ └─┬──┘ ┌─┼──┐ ┌─┼─┐ ┌─┼─┐ ┌─┼──┐ │ElB ├──┤ A ├──┤ B ├──┤ErA │ └─┬──┘ └─┬─┘ └─┬─┘ └─┬──┘ ┌─┼──┐ ┌─┼─┐ ┌─┼─┐ ┌─┼──┐ │Cbl2├──┤EbB├──┤EbA├──┤Cbr2│ └────┘ └───┘ └───┘ └────┘

TNRKit.TRG — Type

mutable struct TRG{E, S, TT<:TensorKit.AbstractTensorMap{E, S, 2, 2}} <: TNRScheme{E, S}

Tensor Renormalization Group

Constructors

TRG(T)

Running the algorithm

run!(::TRG, trunc::TensorKit.TruncationSheme, stop::Stopcrit[, finalizer=default_Finalizer, finalize_beginning=true, verbosity=1])

Each step rescales the lattice by a (linear) factor of √2

verbosity levels

0: No output

1: Print information at start and end of the algorithm

2: Print information at each step

Fields

T::TensorKit.AbstractTensorMap{E, S, 2, 2} where {E, S}: central tensor

References

Levin & Nave Phys. Rev. Letters 99(12) (2007)

TNRKit.c4vCTM — Type

mutable struct c4vCTM{A, S, TT<:TensorKit.AbstractTensorMap{A, S, 0, 4}, TC<:TensorKit.AbstractTensorMap{A, S, 1, 1}, TE<:TensorKit.AbstractTensorMap{A, S, 2, 1}} <: TNRScheme{A, S}

C4v symmetric Corner Transfer Matrix Renormalization Group

Constructors

c4vCTM(T)
c4vCTM(T, [, symmetrize=false])

c4vCTM can be called with a (2,2) tensor (West, South, North, East) with the usual arrow conventions (flipped arrow convention), or with a (0,4) tensor (North, East, South, West) (unflipped arrow convention). The keyword argument symmetrize makes the tensor C4v symmetric when set to true. If symmetrize = false, it checks the symmetry explicitly.

Running the algorithm

run!(::c4vCTM, trunc::TensorKit.TruncationScheme, stop::Stopcrit[, finalize_beginning=true, verbosity=1])

verbosity levels

0: No output

1: Print information at start and end of the algorithm

2: Print information at each step

Fields

T::TensorKit.AbstractTensorMap{A, S, 0, 4} where {A, S}: Central tensor

C::TensorKit.AbstractTensorMap{A, S, 1, 1} where {A, S}: Corner tensor

E::TensorKit.AbstractTensorMap{A, S, 2, 1} where {A, S}: Edge tensor

TNRKit.c6vCTM_triangular — Type

mutable struct c6vCTM_triangular{A, S}

C6v symmetric Corner Transfer Matrix Renormalization Group

Constructors

c6vCTM_triangular(T)
c6vCTM_triangular(T, [, symmetrize=false])

 (120°)     (60°)
    ╲       ╱
     ╲     ╱
      ╲   ╱

(180°)––- T ––-(0°) ╱ ╲ ╱ ╲ ╱ ╲ (240°) (300°)

c6vCTM can be called with a (3, 3) tensor, where the directions are (180°, 240°, 300°, 120°, 60°, 0°) clockwise with respect to the positive x-axis. In the flipped arrow convention, the arrows point from (120°, 60°, 0°) to (180°, 240°, 300°). or with a (0,6) tensor (120°, 60°, 0°, 300°, 240°, 180°) where all arrows point inward (unflipped arrow convention). The keyword argument symmetrize makes the tensor C6v symmetric when set to true. If symmetrize = false, it checks the symmetry explicitly.

Running the algorithm

run!(::c6vCTM, trunc::TensorKit.TruncationScheme, stop::Stopcrit[, finalize_beginning=true, projectors=:twothirds, conditioning=true, verbosity=1])

projectors can either be :twothirds or :full, determining the type of projectors used in the renormalization step. This is based on https://arxiv.org/abs/2510.04907v1. conditioning determines whether to condition the second projector construction. This is based on https://doi.org/10.1103/PhysRevB.98.235148.

verbosity levels

0: No output

1: Print information at start and end of the algorithm

2: Print information at each step

Fields

T::TensorKit.TensorMap{A, S, 0, 6, A1} where {A, S, A1<:DenseVector{A}}

C::TensorKit.TensorMap{A, S, 2, 1, A1} where {A, S, A1<:DenseVector{A}}

Ea::TensorKit.TensorMap{A, S, 2, 1, A1} where {A, S, A1<:DenseVector{A}}

Eb::TensorKit.TensorMap{A, S, 2, 1, A1} where {A, S, A1<:DenseVector{A}}

TNRKit.area_term — Method

The "canonical" normalization constant for loop-TNR tensors, which is the eigenvalue with largest real part of the 2 x 2 transfer matrix.

TNRKit.central_charge — Method

central_charge(scheme::TNRScheme, n::Number)

Get the central charge given the current state of a TNRScheme and the previous normalization factor n

TNRKit.classical_clock — Method

classical_clock(q, β)

Constructs the partition function tensor for the classical clock model with q states and a given inverse temperature β.

TNRKit.classical_ising — Method

classical_ising(β; h)

Constructs the partition function tensor for a 2D square lattice for the classical Ising model with a given inverse temperature β and external magnetic field h.

Examples

    classical_ising() # Default inverse temperature is `ising_βc`
    classical_ising(0.5; h = 1.0) # Custom inverse temperature and magnetic field.

See also: classical_ising_symmetric, classical_ising_symmetric_3D, classical_ising_3D.

TNRKit.classical_ising_3D — Method

classical_ising_3D(β; J)

Constructs the partition function tensor for a 3D cubic lattice for the classical Ising model with a given inverse temperature β and coupling constant J (by default J = 1.0).

Examples

    classical_ising_3D() # Default inverse temperature is `ising_βc_3D`, coupling constant is `J = 1.0`.
    classical_ising_3D(0.5; J = 1.0) # Custom inverse temperature and coupling constant.

See also: classical_ising_symmetric_3D, classical_ising, classical_ising_symmetric.

TNRKit.classical_ising_impurity — Method

classical_ising_impurity(β; h)

Constructs the partition function tensor for a 2D square lattice for the classical Ising model with a given inverse temperature β and external magnetic field h with a magnetisation impurity

Examples

    classical_ising_impurity() # Default inverse temperature is `ising_βc`
    classical_ising_impurity(0.5; h = 1.0) # Custom inverse temperature and magnetic field

Info

When calculating the free energy with free_energy(), set the initial_size keyword argument to 2.0. The initial lattice holds 2 spins.

See also: classical_ising_symmetric, classical_ising_symmetric_3D, classical_ising_3D.

TNRKit.classical_ising_symmetric — Method

classical_ising_symmetric(β)

Constructs the partition function tensor for a symmetric 2D square lattice for the classical Ising model with a given inverse temperature β.

This tensor has explicit ℤ₂ symmetry on each of it spaces.

Examples

    classical_ising_symmetric() # Default inverse temperature is `ising_βc`
    classical_ising_symmetric(0.5) # Custom inverse temperature.

See also: classical_ising, classical_ising_symmetric_3D, classical_ising_3D.

TNRKit.classical_ising_symmetric_3D — Method

classical_ising_symmetric_3D(β)

Constructs the partition function tensor for a symmetric 3D cubic lattice for the classical Ising model with a given inverse temperature β.

This tensor has explicit ℤ₂ symmetry on each of its spaces.

Examples

    classical_ising_symmetric_3D() # Default inverse temperature is `ising_βc_3D`
    classical_ising_symmetric_3D(0.5) # Custom inverse temperature.

See also: classical_ising_3D, classical_ising, classical_ising_symmetric.

TNRKit.classical_ising_triangular — Method

classical_ising_triangular(β)

Constructs the partition function tensor for a 2D triangular lattice for the classical Ising model with a given inverse temperature β.

Examples

    classical_ising_triangular() # Default inverse temperature is `ising_βc_triangular`
    classical_ising_triangular(0.5; h = 1.0) # Custom inverse temperature.

See also: classical_ising_triangular_symmetric.

TNRKit.classical_ising_triangular_symmetric — Method

classical_ising_triangular_symmetric(β)

Constructs the partition function tensor for a symmetric 2D triangular lattice for the classical Ising model with a given inverse temperature β.

This tensor has explicit ℤ₂ symmetry on each of it spaces.

Examples

    classical_ising_triangular_symmetric() # Default inverse temperature is `ising_βc_triangular`
    classical_ising_triangular_symmetric(0.5) # Custom inverse temperature.

See also: classical_ising_triangular.

TNRKit.classical_potts — Method

classical_potts(q, β)

Constructs the partition function tensor for the classical Potts model with q states and a given inverse temperature β.

Examples

    classical_potts(3) # Default inverse temperature is `potts_βc(3)`
    classical_potts(3, 0.5) # Custom inverse temperature.

See also: classical_potts_symmetric, potts_βc.

TNRKit.classical_potts_impurity — Function

classical_potts_impurity(q, β)
classical_potts_impurity(q, β, k1)
classical_potts_impurity(q, β, k1, k2)

Constructs the partition function tensor for a Potts model with q states and a given inverse temperature β with impurities in sectors k1 and k2.

Examples

    classical_potts_impurity(3) # Default inverse temperature is `potts_βc(3)`
    classical_potts_impurity(3, 1, 2) # Custom inverse temperature and impurity sectors.
    classical_potts_impurity(3, 0.5, 2, 3) # Custom inverse temperature and impurity sectors.

See also: classical_potts, potts_βc.

TNRKit.classical_potts_symmetric — Method

classical_potts_symmetric(q, β)

Constructs the partition function tensor for a symmetric Potts model with q states and a given inverse temperature β.

This tensor has explicit ℤq symmetry on each of its spaces.

Examples

    classical_potts_symmetric(3) # Default inverse temperature is `potts_βc(3)`
    classical_potts_symmetric(3, 0.5) # Custom inverse temperature.

See also: classical_potts, potts_βc.

TNRKit.free_energy — Method

free_energy(data, β; scalefactor, initial_size)

Takes a vector of normalization factors data and a given inverse temperature β and computes the free energy.

Info

The scalefactor should be set to the rescaling factor of the area of the tensor network after each iteration of the TNR algorithm.

The initial_size should be set to the intial size of the physical lattice, which is typically 1.0.

TNRKit.gross_neveu_start — Method

gross_neveu_start(μ, m, g)

Constructs the partition function tensor for the Gross-Neveu model with given parameters μ, m, and g.

References

Akiyama et. al. J. Phys.: Condens. Matter 36 (2024) 343002

TNRKit.potts_βc — Method

potts_βc(q)

returns the inverse critical temperature for the classical q-state Potts model on a 2D square lattice.

See also: classical_potts, classical_potts_symmetric.

TNRKit.sixvertex — Function

sixvertex(elt::Type{<:Number} = Float64, symmetry::Type{<:Sector} = Trivial; a = 1.0, b = 1.0, c = 1.0)
sixvertex(symmetry::Type{<:Sector}; kwargs...)

Constructs the partition function tensor for the six-vertex model with a given symmetry type and coupling constants a, b, and c.

Defaults

- elt: Float64
- symmetry: Trivial
- a: 1.0
- b: 1.0
- c: 1.0

Examples

    sixvertex() # Default symmetry is `Trivial`, coupling constants are `a = 1.0`, `b = 1.0`, `c = 1.0`.
    sixvertex(CU1Irrep) # CU1 symmetry with default coupling constants.
    sixvertex(ComplexF64, U1Irrep; a = 2.0, b = 3.0, c = 4.0) # U1 symmetry with custom coupling constants and element type.

Note: The free energy density depends on the boundary conditions.